% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Clustering.R
\name{cluster_run}
\alias{cluster_run}
\title{Clustering algorithm wrapper function}
\usage{
cluster_run(fcs_in, colsToUse, k = 100, xdims = 10, ydims = 10,
  method = "Kmeans", initializer = "kmeans++", debugmode = FALSE)
}
\arguments{
\item{fcs_in}{compensated and transformed FCS in the flowFrame format as input}

\item{colsToUse}{select the channels to use for the clustering: c(7,8,9,10)}

\item{k}{select the number of desired clusters for the Kmeans-type clustering algorithms (default 100)}

\item{xdims}{number of SOM nodes in x direction for FlowSOM algorithm (default 10)}

\item{ydims}{number of SOM nodes in y direction for FlowSOM algorithm (default 10)}

\item{method}{set clustering algorithm from the following choices: 'Kmeans' (default), 'fSOM'}

\item{initializer}{set initializer for Kmeans methods from the following choices: 'optimal_init', 'quantile_init', 'kmeans++' (default), and 'random'; see \href{https://cran.r-project.org/web/packages/ClusterR/ClusterR.pdf}{ClusterR}}
}
\value{
A \code{list} with four items: the first is the clustering result object directly from ClusterR or FlowSOM 
        containing all information as originally constructed in the respective clustering packages, the second is the cluster membership mapping 
        according the the original sequence from the input flowFrame, the third is the centroids in terms of the medianvalues of the columns 
        from the flowFrame based on the membership mapping, and the fourth is the new flowFrame containing just the k (or xdims X ydims) number of centroids.
        This is a wrapper function for clustering methods included in
        \href{https://cran.r-project.org/web/packages/ClusterR/ClusterR.pdf}{ClusterR} and
        \href{https://bioconductor.org/packages/release/bioc/html/FlowSOM.html}{FlowSOM}
        Please see their documentations for more options.
}
\description{
Will generate the clustering results necessary for downstream filtering/gating and plotting
}
\examples{
# Kmeans with random initializer
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), k=500,
                      method="Kmeans", initializer="random")
                      
# Kmeans with kmeans++ initializer
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), k=500)

# FlowSOM
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), xdims=20, ydims=25, method='fSOM')
}
\seealso{
\code{\link{DAFi}}
}
