% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{FCSTransInput}
\alias{FCSTransInput}
\title{Read in raw FCS file and transform it using FCSTrans}
\usage{
FCSTransInput(rawfcs_path, debugmode = FALSE)
}
\arguments{
\item{rawfcs_path}{path to the raw fcs file}
}
\description{
also automatically relabel the columns using marker labels and find colsToUse
}
