% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{ipceil}
\alias{ipceil}
\title{Set output to 0 when input is less than cutoff value}
\usage{
ipceil(x, cutoff = 0, target = 0)
}
\arguments{
\item{x}{input array}

\item{cutoff}{cutoff number}

\item{target}{target number}
}
\description{
Set output to 0 when input is less than cutoff value
}
