% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{ipscatter}
\alias{ipscatter}
\title{immport linear function - convert scatter values to channel output
linear transformation}
\usage{
ipscatter(x, range = 4096, cutoff = 4096, channelrange = 262144)
}
\arguments{
\item{x}{input array}

\item{range}{range number}

\item{cutoff}{cutoff number}

\item{channelrange}{original channel range}
}
\description{
immport linear function - convert scatter values to channel output
linear transformation
}
