% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{iptime}
\alias{iptime}
\title{immport time function - convert time values to channel output
linear transformation}
\usage{
iptime(x, channelrange)
}
\arguments{
\item{x}{input array}

\item{channelrange}{range of channel}
}
\description{
immport time function - convert time values to channel output
linear transformation
}
