% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{scatterTransform}
\alias{scatterTransform}
\title{linear transformation for the scatter channels}
\usage{
scatterTransform(transformationId = "defaultScatterTransform",
  channelrange = 262144, range = 4096, cutoff = 4096, rescale = TRUE)
}
\arguments{
\item{transformationId}{name of transformation}

\item{channelrange}{original channel range}

\item{range}{total range}

\item{cutoff}{cutoff value}

\item{rescale}{whether to rescale or not}
}
\description{
linear transformation for the scatter channels
}
