% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFi_Transforms.R
\name{timeTransform}
\alias{timeTransform}
\title{linear transformation for the time channel}
\usage{
timeTransform(transformationId = "defaultTimeTransform",
  channelrange = 262144, range = 4096, rescale = TRUE)
}
\arguments{
\item{transformationId}{name of transformation}

\item{channelrange}{original channel range}

\item{range}{total range}

\item{rescale}{whether to rescale or not}
}
\description{
linear transformation for the time channel
}
